+++ R/get_grid_bp.R
@@ -15,7 +15,8 @@ get_grid_bp <- function(raster_path, cellsize = 10000){
   mask_polygon <- terra::as.polygons(mask_buffer)
   aoi <- sf::st_as_sf(mask_polygon, quiet = T)
   crs_init <- sf::st_crs(aoi)
-  aoi <- sf::st_multipolygon(lapply(aoi$geometry[2], function(x) x[1]))
+  # aoi <- sf::st_multipolygon(lapply(aoi$geometry[2], function(x) x[1]))
+  aoi <- sf::st_multipolygon(lapply(aoi$geometry, function(x) x[1]))
   aoi <- sf::st_sf(sf::st_sfc(aoi))
   sf::st_crs(aoi) <- crs_init
   aoi_grid <- aoi |> sf::st_make_grid(cellsize = cellsize, square = TRUE)