+++ R/spec_soil_atbd_v2-data.R
@@ -1,9 +1,9 @@
 #' soil reflectance from OSSL database identified for PROSAIL simulations based
-#' on S2 ATBD v3
+#' on S2 ATBD v2
 #'
-#' corresponds to spectral bands, and a selection of 47 soil reflectances
+#' corresponds to spectral bands, and a selection of 7 soil reflectances
 #'
 #' @format A dataframe including 2101 rows and 2 columns
 #'  - *lambda*: wavelength, between 400 nm and 2500 nm
-#'  - *soil_01* to *soil_47*: soil reflectance samples
-"spec_soil_ossl"
+#'  - *soil_01* to *soil_07*: soil reflectance samples
+"spec_soil_atbd_v2"